--- src/main/java/com/conchordance/fretted/FretboardListener.java
@@ -1,5 +1,0 @@
-package com.conchordance.fretted;
-
-public interface FretboardListener {
-	public void fretboardChanged();
-}
