+++ src/main/java/com/conchordance/fretted/FretboardModel.java
@@ -1,140 +1,109 @@
 package com.conchordance.fretted;
 
-import java.util.ArrayList;
 import java.util.LinkedList;
-import java.util.List;
 
 import com.conchordance.fretted.fingering.ChordFingering;
 import com.conchordance.music.Chord;
+import com.conchordance.music.Interval;
 import com.conchordance.music.IntervalicNote;
+import com.conchordance.music.Note;
 
-/**
- * Class to model the fretboard of an instrument. The Fretboard does the brunt of the work of generating chords.
- * Given a chord, it finds all the positions on the neck that are part of the chord, and generates ChordFingerings out of fingerable combinations of those positions.
- * The Fretboard maintains the configuration of capos on the neck, and calculates the chords based on capos.
- *
- */
-public abstract class FretboardModel {
-	public abstract void setChord(Chord c);
+public class FretboardModel {
+
+	public Chord getChord() {
+		return chord;
+	}
+
+	public void setChord(Chord chord) {
+		this.chord = chord;
+		updateInChord();
+	}
 	
-	public abstract Instrument getInstrument();
-	
-	public abstract void setInstrument(Instrument instrument);
+	public Instrument getInstrument() {
+		return instrument;
+	}
+
+	public void setInstrument(Instrument instrument) {
+		this.instrument = instrument;
+		updateInChord();
+	}
+
+	public boolean hasChordNoteAt(int string, int fret) {
+		return fret >= instrument.fretNutPositions[string] && inChord[string][fret] != null;
+	}
 
 	/**
-	 * Determines whether a position on the neck is a note that is part of the chord.
-	 * This includes notes that are not accessible because of a capo, or are out of the fret range.
+	 * Gets the chord note that is at a position on the neck.
+	 *
+	 * This is the note within the Fretboard's current chord, so is not defined for a string/fret that doesn't belong
+	 * in the current chord.
 	 */
-	public abstract boolean hasChordNoteAt(int string, int fret);
-	
-	/**
-	 * Whether a finger can be placed on a position on the fretboard.
-	 * A position is out of range if it is outside of the range specified by the min/max frets, or if it is nut-side of a capo.
-	 */
-	public abstract boolean isInRange(int string, int fret);
+	public IntervalicNote getChordNoteAt(int string, int fret) {
+		return inChord[string][fret];
+	}
 
-    /**
-     * Gets the chord note that is at a position on the neck.
-     *
-     * This is the note within the Fretboard's current chord, so is not defined for a string/fret that doesn't belong
-     * in the current chord
-     */
-    public abstract IntervalicNote getChordNoteAt(int string, int fret);
+	public boolean isInRange(int string, int fret) {
+		return isAtFretNut(string, fret) || fret >= instrument.fretNutPositions[string];
+	}
 
-    /**
-	 * Whether a position is directly covered by a capo.
-	 */
-	public abstract boolean isOnCapo(int string, int fret);
-	
-	/**
-	 * Whether a fret is on a fret nut of a given string.
-	 * Note that for most strings the fret nut is at fret 0, but in some cases (like five string banjo) the fret nut occurs on a higher fret.
-	 */
-	public abstract boolean isAtFretNut(int string, int fret);
-	
-	/**
-	 * Set the minimum fret that can be used when building chords.
-	 * Open notes, whether open to the nut or a capo, are used even if below the minimum fret.
-	 */
-	public abstract void setMinFret(int f);
-	
-	/**
-	 * Gets the minimum fret that can be used when building chords.
-	 */
-	public abstract int getMinFret();
-	
-	/**
-	 * Set the maximum fret that can be used when building chords.
-	 * Capoed notes are used even if above the maximum fret.
-	 */
-	public abstract void setMaxFret(int f);
+	public boolean isAtFretNut(int string, int fret) {
+		return fret == instrument.fretNutPositions[string];
+	}
 
 	/**
-	 * Gets the maximum fret that can be used when building chords.
+	 * The fret number equivalent to "open" for this string.
+	 *
+	 * This takes into account both the position of the fret nut and any capos on that string.
 	 */
-	public abstract int getMaxFret();
+	public int getOpenPosition(int string) {
+		return instrument.fretNutPositions[string];
+	}
 
 	/**
-	 * Gets the set of capos that are on the fretboard.
+	 * Gets the note played when the string is open.
+	 *
+	 * The note may be at the nut or at a capo.
 	 */
-	public abstract List<Capo> getCapos();
+	public IntervalicNote getOpenNote(int string) {
+		return getChordNoteAt(string, getOpenPosition(string));
+	}
 
 	/**
-	 * Adds a capo model to the neck of the instrument, at the location specified by the capo.
+	 * Gets the ChordFingering with the given fret shape on this fretboard.
 	 */
-	public abstract void addCapo(Capo capo);
-	
-	/**
-	 * Creates a new capo for the first available fret and adds it to the instrument.
-	 * If there is no room for a capo, no capo is created or placed.
-	 */
-	public abstract void addCapo();
-	
-	/**
-	 * Invoked when a capo is moved from one place on the fretboard to another.
-	 */
-	public abstract void capoMoved();
-	
-	/**
-	 * Remove a capo from the fretboard.
-	 */
-	public abstract void removeCapo(Capo capo);
+	public ChordFingering getChordFingering(Chord chord, int[] relativeFrets) {
+		int[] absoluteFrets = new int[relativeFrets.length];
+		IntervalicNote[] notes = new IntervalicNote[relativeFrets.length];
 
-    public abstract List<ChordFingering> calculateFingerings();
+		for (int s = 0; s<relativeFrets.length; ++s) {
+			int relativeFret = relativeFrets[s];
+			if (relativeFret != -1) {
+				int absoluteFret = relativeFret == 0 ? getInstrument().fretNutPositions[s] : relativeFret;
+				absoluteFrets[s] = absoluteFret;
+				IntervalicNote note = getChordNoteAt(s, absoluteFret);
+				notes[s] = note;
+			} else {
+				absoluteFrets[s] = -1;
+			}
+		}
 
-    public abstract List<ChordFingering> calculateFingerings(int[] frets);
+		// No finger information is given, so this is blank
+		int[] fingers = new int[relativeFrets.length];
 
-    /**
-     *
-     */
-    public ChordFingering getChordFingering(Chord chord, int[] relativeFrets) {
-        int[] absoluteFrets = new int[relativeFrets.length];
-        IntervalicNote[] notes = new IntervalicNote[relativeFrets.length];
-
-        for (int s = 0; s<relativeFrets.length; ++s) {
-            int relativeFret = relativeFrets[s];
-            if (relativeFret != -1) {
-                int absoluteFret = relativeFret == 0 ? getInstrument().fretNutPositions[s] : relativeFret;
-                absoluteFrets[s] = absoluteFret;
-                IntervalicNote note = getChordNoteAt(s, absoluteFret);
-                notes[s] = note;
-            } else {
-                absoluteFrets[s] = -1;
-            }
-        }
-
-        // No finger information is given, so this is blank
-        int[] fingers = new int[relativeFrets.length];
-
-        return new ChordFingering(chord, absoluteFrets, relativeFrets, fingers, notes, false);
-    }
+		return new ChordFingering(chord, absoluteFrets, relativeFrets, fingers, notes, false);
+	}
 
 	/**
-	 * Produce an array representing each possible combination of strings that could be played open.
-	 * Each combination is represented as a boolean array. Index the array by string number for whether that string is used as part of the combination.
-	 * @param validStrings the strings that can be played open
+	 * All combinations of open strings that can be included in the current chord.
+	 * Each combination is represented by a boolean array that is indexed by string number
 	 */
-	public boolean[][] openStringCombinations(ArrayList<Integer> validStrings) {
+	public boolean[][] openStringCombinations() {
+		LinkedList<Integer> validStrings = new LinkedList<>();
+		for (int string = 0; string<instrument.strings; ++string) {
+			if (hasChordNoteAt(string, getOpenPosition(string)))
+				validStrings.add(string);
+		}
+
 		// combinations = 2^(possible open strings)-1
 		int combos = (int)Math.pow(2, validStrings.size())-1; // no combination corresponding to zero strings (all zeroes in binary)
 
@@ -151,19 +120,48 @@ public abstract class FretboardModel {
 		}
 		return bitPatterns;
 	}
-	
-	public void addFretboardListener(FretboardListener listener) {
-		listeners.add(listener);
+
+	private void updateInChord() {
+		inChord = new IntervalicNote[instrument.strings][instrument.frets+1];
+		for (int string = 0; string<instrument.strings; ++string) {
+			// Iterate up this string starting at the nut (for this string)
+			for (int fret = instrument.fretNutPositions[string]; fret<=instrument.frets; ++fret) {
+				// Search the notes of the chord for a matching pitch
+				for (int i = 0; i<chord.notes.length; ++i) {
+					Note chordNote = chord.notes[i];
+					if (chordNote.halfSteps % 12 == note(string,fret) % 12) {
+						// TODO in rare cases a B# would appear to be a C, or a Cb would appear to be a B# and thus
+						// the octave is incorrect. The note name should be taken into account to get the correct octave
+						int octave = note(string,fret) / 12;
+						Interval interval = chord.intervals[i];
+						inChord[string][fret] = new IntervalicNote(interval, new Note(chordNote.noteName, chordNote.modifier, octave));
+					}
+				}
+			}
+		}
+	}
+
+	/**
+	 * The int-value of the note at a position on the fretboard, based on an absolute fret
+	 */
+	private int note(int string, int fret) {
+		// Subtracting the fret nut position does nothing in the normal case, but for banjo-like instruments
+		// it moves the open note value up to the correct fret (fifth fret for example)
+		return fret + instrument.tuning[string].halfSteps - instrument.fretNutPositions[string];
+	}
+
+	public FretboardModel(Instrument instrument, Chord chord) {
+		this.instrument = instrument;
+		this.chord = chord;
+
+		updateInChord();
 	}
 
 	public FretboardModel() {
-		listeners = new LinkedList<>();
+		this(Instrument.GUITAR, Chord.A_MAJOR);
 	}
-	
-	protected void notifyFretboardChanged() {
-		for (FretboardListener l : listeners)
-			l.fretboardChanged();
-	}
-	
-	private LinkedList<FretboardListener> listeners;
+
+	private Instrument instrument;
+	private Chord chord;
+	private IntervalicNote[][] inChord;
 }
