+++ src/test/java/com/conchordance/fretted/InstrumentTest.java
@@ -6,10 +6,10 @@ import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
 import java.util.Arrays;
-import java.util.HashSet;
 import java.util.LinkedList;
 import java.util.List;
 
+import com.conchordance.fretted.fingering.RecursiveChordFingeringGenerator;
 import com.conchordance.fretted.fingering.list.ChordListModel;
 import com.conchordance.fretted.fingering.validation.ExactShapeValidator;
 import com.conchordance.music.NoteName;
@@ -27,10 +27,11 @@ public class InstrumentTest {
 	@Test
 	public void testAlternateFingerings() {
 		System.out.println("Alternate fingerings:");
-		FretboardModel model = new RecursionBasedFretboardModel(Chord.A_MAJOR);
+		FretboardModel fretboard = new FretboardModel(Instrument.GUITAR, Chord.A_MAJOR);
+
 		int[] aMajorFrets = new int[]{0, 2, 2, 2, 0, 0};
 
-		List<ChordFingering> fingerings = model.calculateFingerings(aMajorFrets);
+		List<ChordFingering> fingerings = new RecursiveChordFingeringGenerator().getChordFingerings(fretboard, aMajorFrets);
 		ChordListModel chords = new ChordListModel();
 		chords.setFilterDuplicateShapes(false);
 		chords.setValidator(new ExactShapeValidator(aMajorFrets));
@@ -49,30 +50,31 @@ public class InstrumentTest {
      */
     @Test
     public void testChordNoteLayout() {
-        FretboardModel model = new RecursionBasedFretboardModel(Chord.A_MAJOR);
+        FretboardModel fretboard = new FretboardModel();
 
         Chord gMajor = new Chord(new Note(NoteName.G, 0), ChordType.MAJOR);
         Chord cMajor = new Chord(new Note(NoteName.C, 0), ChordType.MAJOR);
 
         // guitar
-        model.setInstrument(Instrument.GUITAR);
-        assertEquals("Guitar 1st string open", new Note(NoteName.E, 0, 5), model.getChordNoteAt(0, 0).note);
-        assertEquals("Guitar 1st string octave", new Note(NoteName.E, 0, 6), model.getChordNoteAt(0, 12).note);
-        assertEquals("Guitar 5th string octave", new Note(NoteName.A, 0, 3), model.getChordNoteAt(4, 0).note);
-        assertEquals("Guitar 6th string open", new Note(NoteName.E, 0, 3), model.getChordNoteAt(5, 0).note);
-        assertEquals("Guitar 6th string octave", new Note(NoteName.E, 0, 4), model.getChordNoteAt(5, 12).note);
+        fretboard.setInstrument(Instrument.GUITAR);
+		fretboard.setChord(Chord.A_MAJOR);
+        assertEquals("Guitar 1st string open", new Note(NoteName.E, 0, 5), fretboard.getChordNoteAt(0, 0).note);
+        assertEquals("Guitar 1st string octave", new Note(NoteName.E, 0, 6), fretboard.getChordNoteAt(0, 12).note);
+        assertEquals("Guitar 5th string octave", new Note(NoteName.A, 0, 3), fretboard.getChordNoteAt(4, 0).note);
+        assertEquals("Guitar 6th string open", new Note(NoteName.E, 0, 3), fretboard.getChordNoteAt(5, 0).note);
+        assertEquals("Guitar 6th string octave", new Note(NoteName.E, 0, 4), fretboard.getChordNoteAt(5, 12).note);
 
         // C's on guitar (do the octaves roll over correctly?)
-        model.setChord(cMajor);
-        assertEquals("Guitar 5th 3rd fret", new Note(NoteName.C, 0, 4), model.getChordNoteAt(4, 3).note);
+        fretboard.setChord(cMajor);
+        assertEquals("Guitar 5th 3rd fret", new Note(NoteName.C, 0, 4), fretboard.getChordNoteAt(4, 3).note);
 
         // 5-string banjo (irregular tuning pegs)
-        model.setInstrument(Instrument.BANJO);
-        model.setChord(Chord.A_MAJOR);
-        assertEquals("Banjo 5th string 7th fret", new Note(NoteName.A, 0, 5), model.getChordNoteAt(4, 7).note);
+        fretboard.setInstrument(Instrument.BANJO);
+        fretboard.setChord(Chord.A_MAJOR);
+        assertEquals("Banjo 5th string 7th fret", new Note(NoteName.A, 0, 5), fretboard.getChordNoteAt(4, 7).note);
 
-        model.setChord(gMajor);
-        assertEquals("Banjo 5th string open (5th fret)", new Note(NoteName.G, 0, 5), model.getChordNoteAt(4, 5).note);
+        fretboard.setChord(gMajor);
+        assertEquals("Banjo 5th string open (5th fret)", new Note(NoteName.G, 0, 5), fretboard.getChordNoteAt(4, 5).note);
 
         // guitar, half-capo
 
@@ -87,10 +89,9 @@ public class InstrumentTest {
 	@Test
 	public void openChord() {
 		Instrument openA = new Instrument("Open A", 1, Chord.A_MAJOR.notes[0], Chord.A_MAJOR.notes[0], Chord.A_MAJOR.notes[1], Chord.A_MAJOR.notes[2]);		
-		FretboardModel model = new RecursionBasedFretboardModel(Chord.A_MAJOR);
-		model.setInstrument(openA);
+		FretboardModel fretboard = new FretboardModel(openA, Chord.A_MAJOR);
 		
-		List<ChordFingering> chords = model.calculateFingerings();
+		List<ChordFingering> chords = new RecursiveChordFingeringGenerator().getChordFingerings(fretboard);
 		assertEquals("Only one chord found", 1, chords.size());
 	}
 	
@@ -101,7 +102,7 @@ public class InstrumentTest {
 	 */
 	@Test
 	public void campfireChords() {
-		FretboardModel fretboard = new RecursionBasedFretboardModel(Chord.A_MAJOR);
+		FretboardModel fretboard = new FretboardModel(Instrument.GUITAR, Chord.A_MAJOR);
 
 		// Movable A chords
 		testFingering("A-form A major", new int[] {0, 2, 2, 2, 0, -1}, new Chord(new Note(A, 0), ChordType.MAJOR), fretboard);
@@ -119,11 +120,11 @@ public class InstrumentTest {
 		testFingering("Open G major", new int[] {3, 0, 0, 0, 2, 3}, new Chord(new Note(G, 0), ChordType.MAJOR), fretboard);
 	}
 	
-	private void testFingering(String name, int[] frets, Chord chord, FretboardModel model) {
-		model.setChord(chord);
+	private void testFingering(String name, int[] frets, Chord chord, FretboardModel fretboard) {
+		fretboard.setChord(chord);
 		boolean found = false;
 		
-		List<ChordFingering> chords = model.calculateFingerings();
+		List<ChordFingering> chords = new RecursiveChordFingeringGenerator().getChordFingerings(fretboard);
 		for (ChordFingering fingering : chords) {
 			if (Arrays.equals(frets, fingering.absoluteFrets)) {
 				found = true;
